+++ NEWS.md
@@ -58,6 +58,7 @@ installation.
   Silicon when installed with Homebrew (#240).
 - The `deps-in-desc` hook now points to the hook argument 
   `--allow_private_imports` when the hook fails due to private imports (#254).
+- roxygenize hook is now fully tested (#267).
 - Hook dependency updates are proposed by an automatic monthly pull request 
   to `lorenzwalthert/precommit`. This does not affect users directly (#430).